fix(aces): split image-free helpers out; cast() over replace() type gap
- _compose_realization.py had grown to 511 lines (SonarCloud file-
  length gate, 500 max) after the prior return-count refactor. Splits
  the four pure ADR-048 image-free-partitioning helpers
  (_strip_image_free_published_ports, _image_free_node_addresses,
  _image_free_service_names, _realize_node_subset) into their own
  module, _compose_image_free_realization.py, re-exported from
  _compose_realization.py's __all__ so existing imports (including
  tests/test_mixed_realization_dispatch.py) are unaffected.
- The prior fix for dataclasses.replace()'s return type narrowing to
  DataclassInstance (S5655/S5886) used an explicit variable
  annotation, which SonarCloud's Python analyzer does not treat as a
  narrowing signal; switched to typing.cast(), which it does.
